fix: delete copy-files

+++ 04-copy-directory/index.js
@@ -8,21 +8,23 @@ fs.mkdir(createFolder, { recursive: true }, (err) => {
   if(err) throw err;
 });
 
-fs.readdir(createFolder, (err, filesCopy) => {
-  if (err) throw err;
-  filesCopy.forEach( filesCopyItem => {
-    fs.unlink(path.join(__dirname, 'files-copy', filesCopyItem), (err) => {
-      if (err) throw err;
-    })
-  })
-})
-
 fs.readdir(findFolder, (err, files) => {
   if (err) throw err;
   files.forEach( filesItem => {
     fs.copyFile(path.join(__dirname, 'files', filesItem), path.join(__dirname, 'files-copy', filesItem), (err) => {
       if (err) throw err;
     });
+  })
+
+  fs.readdir(createFolder, (err, filesCopy) => {
+    if (err) throw err;
+    filesCopy.forEach( filesCopyItem => {
+      if(!files.includes(filesCopyItem)) {
+        fs.unlink(path.join(__dirname, 'files-copy', filesCopyItem), (err) => {
+          if (err) throw err;
+        })
+      }
+    })
   })
   console.log(`Files copied successfully`)
 });
